+++ setup/ason_project.sh
@@ -1,0 +1,6 @@
+mkdir -p doc/Logo
+echo "# Change Log(更新日志)" >> ./doc/CHANGELOG.md
+mkdir -p script
+mkdir -p test
+mkdir -p src
+touch README.md